requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
requests==2.32.2
python-dotenv==1.0.0
pydantic==2.4.2 
google-genai==1.1.0
eng_to_ipa==0.0.2
aiohttp==3.9.1